working on product endpoint

+++ src/controllers/product.ts
@@ -1,24 +1,69 @@
-import { NextFunction, Request, Response } from "express";
+import { Request, Response } from "express";
 import { CallbackError } from "mongoose";
-import Category from "../models/Category";
 import Product from "../models/Product";
-import User from "../models/User";
+import { TUser } from "../models/User";
+import Category from "../models/Category";
 
-async function createProduct(req: Request, res: Response, next: NextFunction) {
-  const { id, name, imageUrl, description, categoryId } = req.body;
-  try {
-    const user = await User.isValidUser(id);
-    // const category = await Category.isValidCategory(categoryId, id);
-    res.json(user);
-  } catch (err) {
-    res.json({ user: null });
+interface IResponseWithUser extends Response {
+  locals: { user: TUser | null };
+}
+
+async function createProduct(req: Request, res: IResponseWithUser) {
+  const user = res.locals.user;
+  if (user) {
+    const { name, imageUrl, description, categoryId } = req.body;
+    try {
+      await Category.checkUserCategory(categoryId, user._id);
+      const product = await Product.create({
+        userId: user._id,
+        name,
+        imageUrl,
+        description,
+        categoryId,
+      });
+      res.json({ product });
+    } catch (err) {
+      res.status(400).json(handleErrors(err));
+    }
+  } else {
+    res.status(401).json({
+      error: {
+        message: "User is not authenticated",
+      },
+    });
   }
 }
 
-async function deleteProduct(req: Request, res: Response, next: NextFunction) {
-  const { id } = req.params;
+async function findAllUserProducts(req: Request, res: IResponseWithUser) {
+  const user = res.locals.user;
+  if (user) {
+    try {
+      const products = await Product.find({ userId: user._id });
+      res.json({ products });
+    } catch (err) {
+      res.status(400).json({ products: null });
+    }
+  } else{
+    res.status(401).json({error:"User is not authenticated"})
+  }
+}
 
-  Product.deleteOne({ _id: id }, (err: CallbackError) => {
+async function findProductById(req: Request, res: IResponseWithUser) {
+  const { id } = req.params;
+  const { userId } = req.body;
+  try {
+    const product = await Product.findOne({ _id: id, userId });
+    res.json({ product });
+  } catch (err) {
+    res.status(400).json({ product: null });
+  }
+}
+
+async function deleteProduct(req: Request, res: Response) {
+  const { id } = req.params;
+  const { userId } = req.body;
+
+  Product.deleteOne({ _id: id, userId }, (err: CallbackError) => {
     if (!err) {
       res.json("Product deleted successfully.");
     } else {
@@ -27,26 +72,22 @@ async function deleteProduct(req: Request, res: Response, next: NextFunction) {
   });
 }
 
-function handleModelErrors(err: {
-  message: string;
-  code: number;
-  errors: any;
-}) {
-  let errors: any;
+function handleErrors(err: { message: string; code: number; errors: any }) {
+  let errors: any = {};
 
-  if (err.message === "Invalid username and/or password") {
-    errors.error = err.message;
-  }
-  if (err.code == 11000) {
-    errors.username = "Username is already taken.";
-  }
-  if (err.message.toLowerCase().includes("user validation failed")) {
+  errors.message = err.message;
+  if (err.errors) {
     Object.values(err.errors).forEach((value: any) => {
-      //@ts-ignore
       errors[value.properties.path] = value.properties.message;
     });
   }
+
   return { errors };
 }
 
-export default { createProduct, deleteProduct };
+export default {
+  createProduct,
+  findAllUserProducts,
+  findProductById,
+  deleteProduct,
+};
